PLT-8058 - Code review changes, plus a structure change
