+++ marconi-chain-index/src/Marconi/ChainIndex/Experimental/Api/JsonRpc/Endpoint/Echo.hs
@@ -1,0 +1,26 @@
+{-# LANGUAGE DataKinds #-}
+
+module Marconi.ChainIndex.Experimental.Api.JsonRpc.Endpoint.Echo (
+  RpcEchoMethod,
+  echo,
+) where
+
+import Marconi.ChainIndex.Experimental.Api.Types (HttpServerConfig)
+import Marconi.Core.JsonRpc (ReaderHandler)
+import Network.JsonRpc.Types (JsonRpc, JsonRpcErr)
+
+------------------
+-- Method types --
+------------------
+
+type RpcEchoMethod = JsonRpc "echo" String String String
+
+--------------
+-- Handlers --
+--------------
+
+-- | Echo a message back as a JSON-RPC response. Used for testing the server.
+echo
+  :: String
+  -> ReaderHandler HttpServerConfig (Either (JsonRpcErr String) String)
+echo = return . Right
