+++ marconi-chain-index/src/Marconi/ChainIndex/Experimental/Api/JsonRpc/Endpoint/TargetAddresses.hs
@@ -1,0 +1,37 @@
+{-# LANGUAGE DataKinds #-}
+
+module Marconi.ChainIndex.Experimental.Api.JsonRpc.Endpoint.TargetAddresses (
+  RpcTargetAddressesMethod,
+  getTargetAddressesQueryHandler,
+) where
+
+import Cardano.Api (SerialiseAddress (serialiseAddress))
+import Control.Lens (view)
+import Data.Text (Text)
+import Marconi.ChainIndex.Experimental.Api.Types (HttpServerConfig, configTrackedAddresses)
+import Marconi.Core.JsonRpc (ReaderHandler)
+import Network.JsonRpc.Types (JsonRpc, JsonRpcErr, UnusedRequestParams)
+
+------------------
+-- Method types --
+------------------
+
+type RpcTargetAddressesMethod =
+  JsonRpc
+    "getTargetAddresses"
+    UnusedRequestParams
+    String
+    [Text]
+
+--------------
+-- Handlers --
+--------------
+
+-- | Return the list of TargetAddresses in Bech32 representation.
+getTargetAddressesQueryHandler
+  :: UnusedRequestParams
+  -- ^ Will be an empty string, empty object, or null, as we are ignoring this param, and returning everything
+  -> ReaderHandler HttpServerConfig (Either (JsonRpcErr String) [Text])
+getTargetAddressesQueryHandler _ = do
+  addresses <- view configTrackedAddresses
+  pure $ Right $ serialiseAddress <$> addresses
