+++ marconi-chain-index/src/Marconi/ChainIndex/Experimental/Api/Rest/Endpoint/Params.hs
@@ -1,0 +1,24 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE TypeOperators #-}
+
+module Marconi.ChainIndex.Experimental.Api.Rest.Endpoint.Params where
+
+import Control.Lens (view)
+import Data.Aeson.Types qualified as Aeson
+import Marconi.ChainIndex.Experimental.Api.Types (HttpServerConfig, configParams)
+import Marconi.Core.JsonRpc (ReaderHandler)
+import Servant (Get, JSON, (:>))
+
+------------------
+-- Method types --
+------------------
+
+type GetParams = "params" :> Get '[JSON] Aeson.Value
+
+--------------
+-- Handlers --
+--------------
+
+-- | Returns params given through CLI as a JSON REST response.
+getParamsHandler :: ReaderHandler HttpServerConfig Aeson.Value
+getParamsHandler = view configParams
